+++ spec/dummy/client/webpack.client.base.config.js
@@ -63,12 +63,12 @@ module.exports = {
 
       {
         test: require.resolve('jquery'),
-        use: {
-          loader: 'expose-loader',
-          options: {
-            jQuery: true,
+        use: [
+          {
+            loader: 'expose-loader',
+            options: 'jQuery'
           },
-        },
+        ],
       },
     ],
   },